env: moving main classes to main directory
