+++ src/main/gameIterable.ts
